almost an exact copy of his code... baicly did a merge

+++ devel/AJarn/stable/src/map/mob.h
@@ -18,7 +18,7 @@ struct mob_skill {
 
 struct mob_db {
 	char name[24],jname[24];
-	int lv;
+	short lv;
 	int max_hp,max_sp;
 	int base_exp,job_exp;
 	int atk1,atk2;
