+++ devel/AJarn/stable/src/map/party.h
@@ -24,7 +24,7 @@ int party_member_leaved(int party_id,int account_id,char *name);
 int party_reply_invite(struct map_session_data *sd,int account_id,int flag);
 int party_recv_noinfo(int party_id);
 int party_recv_info(struct party *sp);
-int party_recv_movemap(int party_id,int account_id,char *map,int online,int lv);
+int party_recv_movemap(int party_id,int account_id,char *map,int online,unsigned short lv);
 int party_broken(int party_id);
 int party_optionchanged(int party_id,int account_id,int exp,int item,int flag);
 int party_changeoption(struct map_session_data *sd,int exp,int item);
